Accelerator without finetuning; updates to GPU serving API call
