--- 03_Fine_Tune_Llama2_QLoRA.py
@@ -1,414 +1,0 @@
-# Databricks notebook source
-# MAGIC %md This notebook is available at https://github.com/databricks-industry-solutions/hls-llm-doc-qa
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC # Fine tune Llama-2-7b-hf with QLoRA
-# MAGIC
-# MAGIC [Llama 2](https://huggingface.co/meta-llama) is a collection of pretrained and fine-tuned generative text models ranging in scale from 7 billion to 70 billion parameters. It is trained with 2T tokens and supports context length window upto 4K tokens. [Llama-2-7b-hf](https://huggingface.co/meta-llama/Llama-2-7b-hf) is the 7B pretrained model, converted for the Hugging Face Transformers format.
-# MAGIC
-# MAGIC This is to fine-tune [llama-2-7b-hf](https://huggingface.co/meta-llama/Llama-2-7b-hf) models on the [BI55/MedText](https://huggingface.co/datasets/BI55/MedText) dataset.
-# MAGIC
-# MAGIC Environment for this notebook:
-# MAGIC - Runtime: 13.2 GPU ML Runtime
-# MAGIC - Instance: `g5.8xlarge` on AWS
-# MAGIC
-# MAGIC We will leverage PEFT library from Hugging Face ecosystem, as well as QLoRA for more memory efficient finetuning. Per the [MedText HuggingFace page](https://huggingface.co/datasets/BI55/MedText): 
-# MAGIC "This is a medical diagnosis dataset containing over 1000 top notch textbook quality patient presentations and diagnosis/treatments. The 100 most common diseases and the 30 most common injuries people go to the hospital with, are, among others, fully captured in the dataset, with multiple datapoints for each ranging from mild to complicated to severe. Full list below. The dataset also contains completions about the nature of the AI itself, that it never can replace a doctor and always emphasizes to go to a professional and some nonsensical or doubtful presentations. A model trained on this dataset explicitly tells when it CANNOT answer with confidence or if the presentation is insufficient. This is to prevent hallucinations."
-# MAGIC
-# MAGIC Requirements:
-# MAGIC - To get the access to the model on HuggingFace, please visit the [Meta website](https://ai.meta.com/resources/models-and-libraries/llama-downloads) and accept the license terms and acceptable use policy before submitting this form. Requests will be processed in 1-2 days.
-# MAGIC - Have a Hugging Face account (with the same email address you entered in Meta’s form).
-# MAGIC - Have a Hugging Face token.
-# MAGIC - Visit the page of one of the LLaMA 2 available models (version [7B](https://huggingface.co/meta-llama/Llama-2-7b-hf), [13B](https://huggingface.co/meta-llama/Llama-2-13b-hf) or [70B](https://huggingface.co/meta-llama/Llama-2-70b-hf)), and accept Hugging Face’s license terms and acceptable use policy.
-
-# COMMAND ----------
-
-from huggingface_hub import notebook_login
-
-# Login to Huggingface to get access to the model
-notebook_login()
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Install required packages
-# MAGIC
-# MAGIC Run the cells below to setup and install the required libraries. For our experiment we will need `accelerate`, `peft`, `transformers`, `datasets` and TRL to leverage the recent [`SFTTrainer`](https://huggingface.co/docs/trl/main/en/sft_trainer). We will use `bitsandbytes` to [quantize the base model into 4bit](https://huggingface.co/blog/4bit-transformers-bitsandbytes). We will also install `einops` as it is a requirement to load certain models. [`safetensors`](https://huggingface.co/docs/diffusers/using-diffusers/using_safetensors) is a safer, faster file format for models.
-
-# COMMAND ----------
-
-# MAGIC %run ./util/install-finetune-libraries
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Dataset
-# MAGIC
-# MAGIC We will use the [BI55/MedText](https://huggingface.co/datasets/BI55/MedText) dataset. dataset.
-
-# COMMAND ----------
-
-from datasets import load_dataset
-
-dataset_name = "BI55/MedText"
-dataset = load_dataset(dataset_name, split="train")
-
-# COMMAND ----------
-
-INTRO_BLURB = "Below is a prompt that describes a medical scenario. Write a response that appropriately completes the question."
-INSTRUCTION_KEY = "### Scenario:"
-INPUT_KEY = "Input:"
-RESPONSE_KEY = "### Response:"
-END_KEY = "### End"
-
-PROMPT_NO_INPUT_FORMAT = """{intro}
-
-{instruction_key}
-{instruction}
-
-{response_key}
-{response}
-
-{end_key}""".format(
-  intro=INTRO_BLURB,
-  instruction_key=INSTRUCTION_KEY,
-  instruction="{instruction}",
-  response_key=RESPONSE_KEY,
-  response="{response}",
-  end_key=END_KEY
-)
-
-PROMPT_WITH_INPUT_FORMAT = """{intro}
-
-{instruction_key}
-{instruction}
-
-{input_key}
-{input}
-
-{response_key}
-{response}
-
-{end_key}""".format(
-  intro=INTRO_BLURB,
-  instruction_key=INSTRUCTION_KEY,
-  instruction="{instruction}",
-  input_key=INPUT_KEY,
-  input="{input}",
-  response_key=RESPONSE_KEY,
-  response="{response}",
-  end_key=END_KEY
-)
-
-def apply_prompt_template(examples):
-  instruction = examples["Prompt"]
-  response = examples["Completion"]
-  context = examples.get("context")
-
-  if context:
-    full_prompt = PROMPT_WITH_INPUT_FORMAT.format(instruction=instruction, response=response, input=context)
-  else:
-    full_prompt = PROMPT_NO_INPUT_FORMAT.format(instruction=instruction, response=response)
-  return { "text": full_prompt }
-
-dataset = dataset.map(apply_prompt_template)
-
-# COMMAND ----------
-
-dataset["text"][0]
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Loading the model
-# MAGIC
-# MAGIC In this section we will load [LLaMA-2](https://huggingface.co/meta-llama/Llama-2-7b-hf), quantize it in 4bit and attach LoRA adapters on it.
-
-# COMMAND ----------
-
-import torch
-from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig, AutoTokenizer
-
-model = "meta-llama/Llama-2-7b-chat-hf"
-revision = "0ede8dd71e923db6258295621d817ca8714516d4"
-
-tokenizer = AutoTokenizer.from_pretrained(model, trust_remote_code=True)
-tokenizer.pad_token = tokenizer.eos_token
-tokenizer.add_eos_token = True
-
-bnb_config = BitsAndBytesConfig(
-    load_in_4bit=True,
-    bnb_4bit_quant_type="nf4",
-    bnb_4bit_compute_dtype=torch.float16,
-)
-
-model = AutoModelForCausalLM.from_pretrained(
-    model,
-    quantization_config=bnb_config,
-    device_map="cuda:0",
-    revision=revision,
-    trust_remote_code=True,
-    use_safetensors=True
-)
-model.config.use_cache = False
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC Load the configuration file in order to create the LoRA model. 
-# MAGIC
-# MAGIC According to QLoRA paper, it is important to consider all linear layers in the transformer block for maximum performance. Therefore we will add `dense`, `dense_h_to_4_h` and `dense_4h_to_h` layers in the target modules in addition to the mixed query key value layer.
-
-# COMMAND ----------
-
-from peft import LoraConfig, get_peft_model
-
-lora_alpha = 16 # parameter for scaling
-lora_dropout = 0.1 # dropout probability for layers
-lora_r = 64 # dimension of the updated matrices
-
-peft_config = LoraConfig(
-    lora_alpha=lora_alpha,
-    lora_dropout=lora_dropout,
-    r=lora_r,
-    bias="none",
-    task_type="CAUSAL_LM",
-    target_modules=['q_proj', 'o_proj', 'gate_proj', 'up_proj', 'down_proj', 'k_proj', 'v_proj'] # Choose all linear layers from the model
-)
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Loading the trainer
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC Here we will use the [`SFTTrainer` from TRL library](https://huggingface.co/docs/trl/main/en/sft_trainer) that gives a wrapper around transformers `Trainer` to easily fine-tune models on instruction based datasets using PEFT adapters. Let's first load the training arguments below.
-
-# COMMAND ----------
-
-from transformers import TrainingArguments
-
-output_dir = "/local_disk0/results"
-per_device_train_batch_size = 4
-gradient_accumulation_steps = 4
-optim = "paged_adamw_32bit"
-save_steps = 5
-logging_steps = 10
-learning_rate = 2e-4
-max_grad_norm = 0.3
-max_steps = 5
-warmup_ratio = 0.03
-lr_scheduler_type = "constant"
-
-training_arguments = TrainingArguments(
-    output_dir=output_dir,
-    per_device_train_batch_size=per_device_train_batch_size,
-    gradient_accumulation_steps=gradient_accumulation_steps,
-    optim=optim,
-    save_steps=save_steps,
-    logging_steps=logging_steps,
-    learning_rate=learning_rate,
-    fp16=True,
-    max_grad_norm=max_grad_norm,
-    max_steps=max_steps,
-    warmup_ratio=warmup_ratio,
-    group_by_length=True,
-    lr_scheduler_type=lr_scheduler_type,
-    ddp_find_unused_parameters=False,
-)
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC Then finally pass everthing to the trainer
-
-# COMMAND ----------
-
-from trl import SFTTrainer
-
-max_seq_length = 512 #can make shorter if needed
-
-trainer = SFTTrainer(
-    model=model,
-    train_dataset=dataset,
-    peft_config=peft_config,
-    dataset_text_field="text",
-    max_seq_length=max_seq_length,
-    tokenizer=tokenizer,
-    args=training_arguments,
-)
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC We will also pre-process the model by upcasting the layer norms in float 32 for more stable training
-
-# COMMAND ----------
-
-for name, module in trainer.model.named_modules():
-    if "norm" in name:
-        module = module.to(torch.float32)
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Train the model
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC Now let's train the model! Simply call `trainer.train()`
-
-# COMMAND ----------
-
-trainer.train()
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Save the LORA model
-
-# COMMAND ----------
-
-trainer.save_model("/dbfs/llamav2-7b-MedText-qlora")
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Log the fine tuned model to MLFlow
-
-# COMMAND ----------
-
-import torch
-from peft import PeftModel, PeftConfig
-
-peft_model_id = "/dbfs/llamav2-7b-MedText-qlora"
-config = PeftConfig.from_pretrained(peft_model_id)
-
-from huggingface_hub import snapshot_download
-base_model = "meta-llama/Llama-2-7b-chat-hf"
-revision = "0ede8dd71e923db6258295621d817ca8714516d4"
-# Download the Llama-2-7b-hf model snapshot from huggingface
-snapshot_location = snapshot_download(repo_id=base_model, revision=revision, cache_dir="/local_disk0/.cache/huggingface/", ignore_patterns=["*.bin"])
-
-
-# COMMAND ----------
-
-import mlflow
-class LLAMAQLORA(mlflow.pyfunc.PythonModel):
-  def load_context(self, context):
-    self.tokenizer = AutoTokenizer.from_pretrained(context.artifacts['repository'])
-    self.tokenizer.pad_token = tokenizer.eos_token
-    config = PeftConfig.from_pretrained(context.artifacts['lora'])
-    base_model = AutoModelForCausalLM.from_pretrained(
-      context.artifacts['repository'], 
-      return_dict=True, 
-      load_in_4bit=True, 
-      device_map={"":0},
-      trust_remote_code=True,
-      use_safetensors=True
-    )
-    self.model = PeftModel.from_pretrained(base_model, context.artifacts['lora'])
-  
-  def predict(self, context, model_input):
-    prompt = model_input["prompt"][0]
-    temperature = model_input.get("temperature", [1.0])[0]
-    max_tokens = model_input.get("max_tokens", [100])[0]
-    batch = self.tokenizer(prompt, padding=True, truncation=True,return_tensors='pt').to('cuda')
-    with torch.cuda.amp.autocast():
-      output_tokens = self.model.generate(
-          input_ids = batch.input_ids, 
-          max_new_tokens=max_tokens,
-          temperature=temperature,
-          top_p=0.7,
-          num_return_sequences=1,
-          pad_token_id=tokenizer.eos_token_id,
-          eos_token_id=tokenizer.eos_token_id,
-      )
-    generated_text = self.tokenizer.decode(output_tokens[0], skip_special_tokens=True)
-
-    return generated_text
-
-# COMMAND ----------
-
-from mlflow.models.signature import ModelSignature
-from mlflow.types import DataType, Schema, ColSpec
-import pandas as pd
-import mlflow
-
-# Define input and output schema
-input_schema = Schema([
-    ColSpec(DataType.string, "prompt"), 
-    ColSpec(DataType.double, "temperature"), 
-    ColSpec(DataType.long, "max_tokens")])
-output_schema = Schema([ColSpec(DataType.string)])
-signature = ModelSignature(inputs=input_schema, outputs=output_schema)
-
-# Define input example
-input_example=pd.DataFrame({
-            "prompt":["what is ML?"], 
-            "temperature": [0.5],
-            "max_tokens": [100]})
-
-with mlflow.start_run() as run:  
-    mlflow.pyfunc.log_model(
-        "Llama-2-7b-MedText",
-        python_model=LLAMAQLORA(),
-        artifacts={'repository' : snapshot_location, "lora": peft_model_id},
-        pip_requirements=["torch", "transformers", "accelerate", "einops", "loralib", "safetensors", "bitsandbytes", "peft"],
-        input_example=pd.DataFrame({"prompt":["what is cystic fibrosis?"], "temperature": [0.1],"max_tokens": [150]}),
-        signature=signature
-    )
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC Run model inference with the model logged in MLFlow.
-
-# COMMAND ----------
-
-import mlflow
-import pandas as pd
-
-
-prompt = """Below is an instruction that describes a task. Write a response that appropriately completes the request.
-### Instruction:
-What is the proper course of response if you are diagnosed with coronaviraus?
-
-### Response: """
-# Load model as a PyFuncModel.
-run_id = run.info.run_id
-logged_model = f"runs:/{run_id}/Llama-2-7b-MedText"
-
-loaded_model = mlflow.pyfunc.load_model(logged_model)
-
-text_example=pd.DataFrame({
-            "prompt":[prompt], 
-            "temperature": [0.1],
-            "max_tokens": [150]})
-
-# Predict on a Pandas DataFrame.
-loaded_model.predict(text_example)
-
-# COMMAND ----------
-
-displayHTML(loaded_model.predict(text_example))
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ### Register the model
-
-# COMMAND ----------
-
-# Register model in MLflow Model Registry
-# This may take about 6 minutes to complete
-result = mlflow.register_model(
-    "runs:/"+run.info.run_id+"/Llama-2-7b-MedText",
-    name="Llama-2-7b-MedText-QLoRa",
-    await_registration_for=1000,
-)
